fix error when anidb returns error
